simplify babel configs
