+++ catalogue/webapp/package.json
@@ -13,11 +13,8 @@
   },
   "dependencies": {
     "@weco/common": "../../common",
-    "@weco/next-plugin-transpile-modules": "^0.0.2",
-    "@zeit/next-bundle-analyzer": "^0.1.1",
     "autoprefixer": "^8.3.0",
     "babel-plugin-module-resolver": "^3.1.1",
-    "babel-plugin-wrap-in-js": "^1.1.1",
     "cookies": "^0.7.1",
     "css-loader": "^0.28.11",
     "isomorphic-unfetch": "^2.0.0",
