+++ common/.babelrc
@@ -4,13 +4,10 @@
       "targets": {
         "node": "current"
       }
-    }]
+    }],
+    "next/babel"
   ],
   "plugins": [
-    "transform-class-properties",
-    "transform-object-rest-spread",
-    "transform-react-jsx",
-    "react-require",
     "transform-flow-strip-types"
   ]
 }
