+++ common/package.json
@@ -6,14 +6,9 @@
   },
   "license": "MIT",
   "dependencies": {
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-module-resolver": "^3.0.0",
-    "babel-plugin-react-require": "^3.0.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
+    "@weco/next-plugin-transpile-modules": "^0.0.2",
+    "@zeit/next-bundle-analyzer": "^0.1.1",
     "babel-plugin-transform-flow-strip-types": "^6.22.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-plugin-transform-react-jsx": "^6.24.1",
-    "babel-plugin-wrap-in-js": "^1.1.1",
     "babel-preset-env": "^1.6.1",
     "cookies": "^0.7.1",
     "fontfaceobserver": "^2.0.13",
@@ -25,7 +20,7 @@
     "lscache": "^1.1.0",
     "moment": "^2.20.1",
     "moment-timezone": "^0.5.14",
-    "next": "^5.0.0",
+    "next": "^5.1.0",
     "openseadragon": "^2.3.1",
     "prismic-dom": "^2.0.6",
     "prismic-javascript": "^1.5.0-beta.2",
