+++ whats_on/app/package.json
@@ -9,9 +9,7 @@
   "dependencies": {
     "@weco/common": "../../common",
     "babel-cli": "^6.26.0",
-    "babel-jest": "^22.4.4",
     "babel-loader": "^7.1.2",
-    "babel-plugin-flow-runtime": "^0.15.0",
     "babel-plugin-module-resolver": "^3.0.0",
     "babel-plugin-transform-flow-strip-types": "^6.22.0",
     "babel-preset-env": "^1.6.1",
